Use standard C99 type size_t and POSIX ssize_t
All occurences of `unsigned int` for array index or length are replaced
with `size_t`, as this type is more suited for these uses. (moreover it
is a wider type, so the arrays can contain even *more* elements!)
All occurences of `int` as an index or error value (usually `-1`) are
replaced with `ssize_t`, the POSIX-defined equivalent signed type --
actually, it is only ensured to contain *at least* integers from the
range `[-1, 2^15 - 1]`, but it is usually the signed counterpart of
`size_t`.

+++ inc/array.h
@@ -29,6 +29,8 @@
 
 
 #include "cods.h" /* for function attrs, data */
+#include <stddef.h> /* for size_t */
+#include <unistd.h> /* for ssize_t */
 
 
 
@@ -59,7 +61,7 @@ typedef struct array Array;
  * \return A new instance of \a Array, with \a size elements slots, or \c NULL
  *         if an error occurred.
  */
-CODS_CTOR Array *a_new(unsigned int size);
+CODS_CTOR Array *a_new(size_t size);
 
 
 /**
@@ -77,7 +79,7 @@ CODS_MEMBER void a_free(Array *self);
  *
  * \return The number of elements in the array.
  */
-CODS_MEMBER unsigned int a_size(const Array *self) CODS_PURE;
+CODS_MEMBER size_t a_size(const Array *self) CODS_PURE;
 
 
 /**
@@ -93,7 +95,7 @@ CODS_MEMBER unsigned int a_size(const Array *self) CODS_PURE;
  *
  * \return The \a index 'th element in the array, or \c NULL.
  */
-CODS_MEMBER data *a_get(const Array *self, unsigned int index) CODS_PURE;
+CODS_MEMBER data *a_get(const Array *self, size_t index) CODS_PURE;
 
 
 /**
@@ -105,7 +107,7 @@ CODS_MEMBER data *a_get(const Array *self, unsigned int index) CODS_PURE;
  * \param[in]     index   The index at which to update the element
  * \param[in]     newitem The new element to set
  */
-CODS_MEMBER void a_set(Array *self, unsigned int index, data *newitem)
+CODS_MEMBER void a_set(Array *self, size_t index, data *newitem)
 CODS_NOTNULL(3);
 
 
@@ -121,7 +123,7 @@ CODS_NOTNULL(3);
  * \param[in]     newitem The element to add
  * \return The new size of the array or \c -1.
  */
-CODS_MEMBER int a_add(Array *self, unsigned int index, data *newitem)
+CODS_MEMBER ssize_t a_add(Array *self, size_t index, data *newitem)
 CODS_NOTNULL(3);
 
 /**
@@ -138,7 +140,7 @@ CODS_NOTNULL(3);
  * \sa a_add
  */
 CODS_MEMBER CODS_INLINE CODS_NOTNULL(2)
-int a_append(Array *self, data *item) {
+ssize_t a_append(Array *self, data *item) {
 	return a_add(self, a_size(self), item);
 }
 
@@ -154,6 +156,6 @@ int a_append(Array *self, data *item) {
  * \return The element just removed, or \c NULL if an error occurred (\a index
  *         is invalid).
  */
-CODS_MEMBER data *a_drop(Array *self, unsigned int index);
+CODS_MEMBER data *a_drop(Array *self, size_t index);
 
 #endif /* CODS_ARRAY_H */
